requirements.txt
selenium==4.18.1
behave==1.2.6
webdriver-manager==4.0.1
Pillow==10.2.0
allure-behave==2.13.5
python-dotenv==1.0.1
faker==23.2.1
